pyproject.toml
[project]
name = "ky-education-kpi-pipeline"
version = "0.1.0"
description = "ETL pipelines to normalize Kentucky Department of Education data for multi-year KPI reporting"
requires-python = ">=3.11"
authors = [{name="Scott Horn", email="scott@kyopengov.org"}]

dependencies = [
    "pandas",
    "pydantic",
    "ruamel-yaml",
    "pytest",
    "tabulate",
    "requests"
]


[tool.setuptools.packages.find]
where = ["."]
include = ["etl*"]
exclude = ["data*", "notes*", "config*", "tests*"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88